address comments

+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/StreamsPartitionAssignor.java
@@ -256,7 +256,7 @@ public class StreamsPartitionAssignor implements ConsumerPartitionAssignor, Conf
             userEndPoint,
             taskManager.getTaskOffsetSums(),
             uniqueField,
-            (byte) assignmentErrorCode.get()
+            assignmentErrorCode.get()
         ).encode();
     }
